ExpressionDTO moved into Expression Endpoint
