+++ src/RegExLib.Web/Endpoints/Expressions/Create.CreateExpressionCommand.cs
@@ -1,7 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using AutoMapper;
 using RegExLib.Core.Entities;
-using RegExLib.Web.ApiModels;
 
 namespace RegExLib.Web.Endpoints.Expressions
 {
