+++ src/RegExLib.Web/Endpoints/Expressions/List.PagedExpressionResult.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using RegExLib.Web.ApiModels;
 
 namespace RegExLib.Web.Endpoints.Expressions
 {
