+++ src/RegExLib.Web/Startup.cs
@@ -12,8 +12,6 @@ using RegExLib.Infrastructure.Identity;
 using Microsoft.AspNetCore.Identity;
 using System;
 using AutoMapper;
-using AutoMapper.Configuration;
-using RegExLib.Web.Endpoints.Expressions;
 using IConfiguration = Microsoft.Extensions.Configuration.IConfiguration;
 
 namespace RegExLib.Web
