+++ tests/RegExLib.FunctionalTests/Api/Expressions/ApiExpressionCreate.cs
@@ -4,9 +4,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.TestHost;
 using Newtonsoft.Json;
-using RegExLib.Web.ApiModels;
 using RegExLib.Web.Endpoints.Expressions;
-using RegExLib.Web.Seeds;
 using Xunit;
 
 namespace RegExLib.FunctionalTests.Api.expressions
